# staffler/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Staffler / DPS-Service API",
    "version": "0.0.671-SNAPSHOT (commit 1fc6cd3)",
    "description": "Reconstructed OpenAPI spec from `wlnob/dps-service` controllers. Scope: REST endpoints exposed by dps-service Spring Boot 3.2 monolith. Inlined real schemas as of 2026-05-09 from sources/dps-service-dtos.md (priority DTOs only; non-priority types may still be x-stub). Source commit: 1fc6cd30d62ec3bba51285585483a524a22f4238.",
    "contact": {
      "name": "BOEMM Staffler team"
    }
  },
  "servers": [
    {
      "url": "https://gw.qa.dps.boemm.eu/v1/dps-api",
      "description": "QA"
    },
    {
      "url": "https://gw.dev.dps.boemm.eu/v1/dps-api",
      "description": "DEV"
    },
    {
      "url": "https://gw.myplanning.digitalpayrollservices.be/v1/dps-api",
      "description": "PROD"
    }
  ],
  "components": {
    "securitySchemes": {
      "BoemmSkey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-boemm-skey",
        "description": "Opaque session key obtained via /publicapi/companies/users/login or /publicapi/employees/users/login. Resolved by API Gateway authorizer to a Cognito JWT."
      },
      "BoemmInternal": {
        "type": "apiKey",
        "in": "header",
        "name": "x-boemm-skey",
        "description": "Internal cron endpoints. Externally not reachable; documented for completeness."
      }
    },
    "schemas": {
      "PageWebDto": {
        "type": "object",
        "description": "Custom paged envelope (NOT Spring's default Page).",
        "properties": {
          "content": {
            "type": "array",
            "items": {}
          },
          "totalElements": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "numberOfElements": {
            "type": "integer"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "GeneralCompanyCoefficientsWebDto": {
        "type": "object",
        "properties": {
          "coefficientTravelAllowance": {
            "type": "number"
          },
          "dimonaCost": {
            "type": "number"
          },
          "dimonaAddon": {
            "type": "number"
          },
          "coefficientMealVouchers": {
            "type": "number"
          },
          "coefficientEcoVouchers": {
            "type": "number"
          },
          "defaultTaxRate": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          }
        }
      },
      "CompanyCoefficientsPerStatuteWebDto": {
        "type": "object",
        "description": "Same DTO reused for both holiday and general coefficient maps. Each field is the per-statute coefficient.",
        "properties": {
          "coefficientWhiteCollar": {
            "type": "number"
          },
          "coefficientBlueCollar": {
            "type": "number"
          },
          "coefficientWhiteCollarJobStudent": {
            "type": "number"
          },
          "coefficientBlueCollarJobStudent": {
            "type": "number"
          },
          "coefficientFlextimeWhiteCollar": {
            "type": "number"
          },
          "coefficientFlextimeBlueCollar": {
            "type": "number"
          },
          "coefficientWhiteCollarStudentWorker": {
            "type": "number"
          },
          "coefficientBlueCollarStudentWorker": {
            "type": "number"
          },
          "coefficientExtra": {
            "type": "number"
          },
          "coefficientSeasonalWorker": {
            "type": "number"
          },
          "coefficientConstructionWorker": {
            "type": "number"
          },
          "coefficientConstructionJobStudent": {
            "type": "number"
          }
        }
      },
...
      "CompanyCreateResultWebDto": {
        "type": "object",
        "description": "Returned by POST /api/companies/{vat}.",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/StatusWebDto"
          }
        }
      },
...
      "CompanyWebDto": {
        "type": "object",
        "description": "Canonical company resource: same shape on read (GET /api/companies/{companyId}) and write (PUT /api/companies/{uuid}). No separate Create/Update DTOs.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "nickName": {
            "type": "string",
            "nullable": true
          },
          "vat": {
            "type": "string"
          },
          "vatCountryCode": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/AddressWebDto"
          },
          "paritairComites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DictionaryItemWebDto"
            }
          },
          "officeCode": {
            "type": "string",
            "nullable": true
          },
          "communication": {
            "$ref": "#/components/schemas/CommunicationWebDto"
          },
          "personalContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonalContactWebDto"
            }
          },
          "blockingReason": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          },
          "blockedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "blockingExtraInfo": {
            "type": "string",
            "nullable": true
          },
          "blockedBy": {
            "$ref": "#/components/schemas/ConsultantWebDto"
          },
          "status": {
            "$ref": "#/components/schemas/CompanyStatusDto"
          },
          "revenueConsultant": {
            "$ref": "#/components/schemas/ConsultantWebDto"
          },
          "coefficients": {
            "$ref": "#/components/schemas/GeneralCompanyCoefficientsWebDto"
          },
          "holidayCoefficientsPerStatute": {
            "$ref": "#/components/schemas/CompanyCoefficientsPerStatuteWebDto"
          },
          "coefficientsPerStatute": {
            "$ref": "#/components/schemas/CompanyCoefficientsPerStatuteWebDto"
          },
          "companyInvoiceInfo": {
            "$ref": "#/components/schemas/CompanyInvoiceInfoWebDto"
          },
          "travelAllowance": {
            "$ref": "#/components/schemas/GeneralCompanyTravelAllowanceWebDto"
          },
          "mealVoucher": {
            "$ref": "#/components/schemas/GeneralCompanyMealVoucherWebDto"
          },
          "externalService": {
            "type": "string",
            "nullable": true
          },
          "externalServiceRegNumber": {
            "type": "string",
            "nullable": true
          },
          "isOnboarded": {
            "type": "boolean",
            "nullable": true
          },
          "isTimeRegistrationEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "isGroupsEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "isActualsEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "actualsBlockEnabled": {
            "type": "boolean",
            "default": true
          },
          "presumedStartDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "socialSecurityCategory": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          }
        }
      },
      "AddressWebDto": {
        "type": "object",
        "description": "Shared by company, contract, employee, invitation. NOT the same as boemm-core AddressDTO.",
        "properties": {
          "streetNumber": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "nullable": true
          },
          "bus": {
            "type": "string",
            "nullable": true
          },
          "formattedAddress": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CommunicationWebDto": {
        "type": "object",
        "description": "Embedded in CompanyWebDto.",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "invoicePhoneNumber": {
            "type": "string",
            "nullable": true
          },
          "selfServiceEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eremindersEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "einvoicesEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          }
        }
      },
      "PersonalContactWebDto": {
        "type": "object",
        "description": "Embedded as a list on CompanyWebDto.personalContacts.",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ContractWebDto": {
        "type": "object",
        "description": "Both request (POST/PUT) and response on ContractController. Single canonical shape; no separate Create/Update DTO. Carries embedded result holder for batch responses. Time fields use HH:mm format inside timetable.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "allocationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ContractStatusWebDto"
          },
          "previousStatus": {
            "$ref": "#/components/schemas/ContractStatusWebDto"
          },
          "dateFrom": {
            "type": "string",
            "format": "date"
          },
          "dateTo": {
            "type": "string",
            "format": "date"
          },
          "timetable": {
            "$ref": "#/components/schemas/ContractTimetableWebDto"
          },
          "flexibleTimetable": {
            "type": "boolean",
            "nullable": true,
            "description": "TODO NULL no scope per source comment."
          },
          "officeCode": {
            "type": "string",
            "nullable": true
          },
          "consultant": {
            "$ref": "#/components/schemas/ConsultantWebDto"
          },
          "revenueOfficeCode": {
            "type": "string",
            "nullable": true
          },
          "revenueConsultant": {
            "$ref": "#/components/schemas/ConsultantWebDto"
          },
          "wageHour": {
            "type": "number"
          },
          "compensationHours": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          },
          "invoiceEcoWeekly": {
            "type": "boolean",
            "nullable": true
          },
          "mealVoucher": {
            "$ref": "#/components/schemas/ContractWebDto_ContractEmployeeWageMealVoucherWebDto"
          },
          "travelAllowance": {
            "$ref": "#/components/schemas/ContractWebDto_ContractEmployeeWageTravelAllowanceWebDto"
          },
          "invoicing": {
            "$ref": "#/components/schemas/ContractInvoicingWebDto"
          },
          "employmentAddress": {
            "$ref": "#/components/schemas/AddressWebDto"
          },
          "paritairComite": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          },
          "statute": {
            "$ref": "#/components/schemas/StatuteItemWebDto"
          },
          "reason": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          },
          "employeeHoursPerWeek": {
            "type": "number",
            "nullable": true
          },
          "companyHoursPerWeek": {
            "type": "number",
            "nullable": true
          },
          "cancelReason": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          },
          "cancelExtraInfo": {
            "type": "string",
            "nullable": true
          },
          "isLate": {
            "type": "boolean",
            "nullable": true
          },
          "sourceType": {
            "$ref": "#/components/schemas/ContractSourceWebDto"
          },
          "shiftTemplate": {
            "$ref": "#/components/schemas/ShiftTemplateWebDto"
          },
          "mutualAgreementContractCancellation": {
            "$ref": "#/components/schemas/MutualAgreementContractCancellationWebDto"
          },
          "result": {
            "$ref": "#/components/schemas/ContractBatchCreationResultWebDto",
            "description": "Populated only on POST /batch responses."
          },
          "socialSecurityCategory": {
            "$ref": "#/components/schemas/DictionaryItemWebDto"
          }
        }
      },
...
      "ConsultantWebDto": {
        "type": "object",
        "description": "Embedded across company / contract / wage DTOs.",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
...
      "CompanyStatusDto": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "BLOCKED",
          "PROCESSING"
        ]
      },
...
      "DictionaryItemWebDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "description": "Generic dictionary item shape from boemm-core. Two fields only."
      },
...
    }
  },
...
  "paths": {
...
    "/api/companies/{companyId}": {
      "get": {
        "operationId": "getCompany",
        "summary": "Get company by id",
        "tags": [
          "company"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWebDto",
                  "x-original-name": "CompanyWebDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-preauthorize": "COMPANY_VIEW",
        "x-controller": "CompanyController",
        "x-module": "company"
      }
    },
...
    "/api/companies/{vat}": {
      "post": {
        "operationId": "createCompany",
        "summary": "Onboard new company by VAT (Credit Safe lookup)",
        "tags": [
          "company"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyCreateResultWebDto",
                  "x-original-name": "CompanyCreateResultWebDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "vat",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-preauthorize": "COMPANY_ONBOARDING",
        "x-controller": "CompanyController",
        "x-module": "company"
      }
    },
...
    "/api/companies/{uuid}": {
      "put": {
        "operationId": "updateCompany",
        "summary": "Update full Company",
        "tags": [
          "company"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWebDto",
                  "x-original-name": "CompanyWebDto"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyWebDto",
                "x-original-name": "CompanyWebDto"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-preauthorize": "checkCompanyUpdate",
        "x-controller": "CompanyController",
        "x-module": "company"
      }
    },
...
    "/api/contracts": {
      "post": {
        "operationId": "createContract",
        "summary": "Create a new contract",
        "tags": [
          "contract"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractWebDto",
                  "x-original-name": "ContractWebDto"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractWebDto",
                "x-original-name": "ContractWebDto"
              }
            }
          }
        },
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-preauthorize": "checkContract",
        "x-controller": "ContractController",
        "x-module": "contract"
      },
      "get": {
        "operationId": "getContracts",
        "summary": "Search contracts (paged ContractBaseWebDto)",
        "tags": [
          "contract"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageWebDto",
                  "x-page-of": "ContractBaseWebDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeStartDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeEndDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-controller": "ContractController",
        "x-module": "contract"
      }
    },
...
    "/api/employees": {
      "get": {
        "operationId": "getCompanyEmployees",
        "summary": "Search employees in a company pool",
        "tags": [
          "employee"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageWebDto",
                  "x-page-of": "EmployeeWebDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nameLike",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "baseView",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasContractFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasContractUntil",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actualFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actualUntil",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actualsStatuses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BoemmSkey": []
          }
        ],
        "x-preauthorize": "EMPLOYEE_VIEW",
        "x-controller": "EmployeeController",
        "x-module": "employee"
      }
    },
...
  },
...
}